+++ src/app/auth/component/signup-template/signup-template.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-signup-template',
-  imports: [],
-  templateUrl: './signup-template.component.html',
-  styleUrl: './signup-template.component.scss'
-})
-export class SignupTemplateComponent {}