+++ src/app/auth/interface/auth-facade.ts
@@ -3,4 +3,5 @@ import { Observable } from 'rxjs';
 export interface IAuthFacade {
   login(username: string, password: string): void;
   logout(): void;
+  signup(username: string, password: string, confirmPassword: string): void;
 }