+++ src/app/shared/constant/navigation-route.const.ts
@@ -4,11 +4,12 @@ export const NavigationRoute = {
     USER_ACTIVATION: 'activate'
   },
   FEATURE: {
+    DASHBOARD: 'dashboard',
     MANAGE_USER: generateRoutes('manage-user')
   }
 };
 
-// function to generate path for pages
+// function to generate path for pages of CRUD for uniformity
 function generateRoutes(basePath: string) {
   return {
     LIST: basePath,